add test of documentation-vs-response validity
In particular responses should never contain
undocumented fields.  In this instance as
noted by gygaxis@zingolabs.com, a field exists
in the actual response that's not in the
help output.

This change also incorporates a test fixture
which should make writing tests require less
boilerplate.

Notes:

  * the class of tests that validates the
documentation will eventually be it's own
mod and likely will be in the tests directory
as it's properly integration testing
  * the machinery implemented concretely here
can (and soon will) be easily abstracted to
test more RPCs than just getinfo (which is
serving as out POC RPC.
  * the overall design is to have many small
specific tests to ease the identification of
specific kinds of failures (as opposed to
"omnibus" tests with many assertions that
cover logically disparate aspects of the code
under test (obviously this is a continuum)